+++ src/components/editor/scroll/utils.ts
@@ -1,4 +1,4 @@
-import { LineMarkerPosition } from '../../markdown-renderer/utils/types'
+import { LineMarkerPosition } from '../../markdown-renderer/types'
 
 export const findLineMarks = (lineMarks: LineMarkerPosition[], lineNumber: number): { lastMarkBefore: LineMarkerPosition | undefined, firstMarkAfter: LineMarkerPosition | undefined } => {
   let lastMarkBefore
