+++ src/components/markdown-renderer/full-markdown-renderer.tsx
@@ -18,7 +18,7 @@ import { InternalLink } from '../common/links/internal-link'
 import { ShowIf } from '../common/show-if/show-if'
 import { slugify } from '../editor/table-of-contents/table-of-contents'
 import { RawYAMLMetadata, YAMLMetaData } from '../editor/yaml-metadata/yaml-metadata'
-import { AdditionalMarkdownRendererProps, BasicMarkdownRenderer } from './basic-markdown-renderer'
+import { BasicMarkdownRenderer } from './basic-markdown-renderer'
 import { createRenderContainer, validAlertLevels } from './markdown-it-plugins/alert-container'
 import { firstHeaderExtractor } from './markdown-it-plugins/first-header-extractor'
 import { highlightedCode } from './markdown-it-plugins/highlighted-code'
@@ -55,7 +55,7 @@ import { TaskListReplacer } from './replace-components/task-list/task-list-repla
 import { VimeoReplacer } from './replace-components/vimeo/vimeo-replacer'
 import { YoutubeReplacer } from './replace-components/youtube/youtube-replacer'
 import { useCalculateLineMarkerPosition } from './utils/calculate-line-marker-positions'
-import { LineMarkerPosition } from './utils/types'
+import { AdditionalMarkdownRendererProps, LineMarkerPosition } from './types'
 import { usePostMetaDataOnChange } from './utils/use-post-meta-data-on-change'
 import { usePostTocAstOnChange } from './utils/use-post-toc-ast-on-change'
